requirements.txt
streamlit>=1.38
pandas>=2.2
plotly>=5.24
python-docx>=1.1
pypdf>=5.0
openpyxl>=3.1
